Function interbtc_parachain::service::start_node_impl

async fn start_node_impl<RB, RuntimeApi, Executor, BIC>(
    parachain_config: Configuration, 
    polkadot_config: Configuration, 
    id: ParaId, 
    _rpc_ext_builder: RB, 
    build_consensus: BIC
) -> Result<(TaskManager, Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>)> where
    RB: Fn(Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>) -> Result<IoHandler<Metadata>, Error> + Send + 'static,
    RuntimeApi: ConstructRuntimeApi<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>> + Send + Sync + 'static,
    RuntimeApi::RuntimeApi: RuntimeApiCollection<StateBackend = StateBackendFor<TFullBackend<Block>, Block>>,
    RuntimeApi::RuntimeApi: AuraApi<Block, AuraId>,
    Executor: NativeExecutionDispatch + 'static,
    BIC: FnOnce(Arc<TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>, Option<&Registry>, Option<TelemetryHandle>, &TaskManager, Arc<dyn RelayChainInterface>, Arc<FullPool<Block, TFullClient<Block, RuntimeApi, NativeElseWasmExecutor<Executor>>>>, Arc<NetworkService<Block, Hash>>, SyncCryptoStorePtr, bool) -> Result<Box<dyn ParachainConsensus<Block>>, Error>, 

Start a node with the given parachain Configuration and relay chain Configuration.

This is the actual implementation that is abstract over the executor and the runtime api.